lab6_project/dynamic/server/requirements.txt
zeroc-ice
icepy
